+++ src/sreg/models/__init__.py
@@ -1,5 +1,6 @@
 """Data contracts for SREG."""
 
+from sreg.models.case_plan import CasePlan, EvalQuestionPlan
 from sreg.models.dag_spec import DAGNodeSpec, DAGSpec
 from sreg.models.episode import Action, ActionType, Episode, Observation, StepResult
 from sreg.models.research_problem import AvailableAction, DataAsset, ResearchProblem
@@ -13,12 +14,14 @@ __all__ = [
     "ActionType",
     "AvailableAction",
     "CPD",
+    "CasePlan",
     "DAGNodeSpec",
     "DAGSpec",
     "DataAsset",
     "DifficultyProfile",
     "Edge",
     "Episode",
+    "EvalQuestionPlan",
     "Node",
     "NodeType",
     "Observation",